@@ -1,149 +1,84 @@
 name: pdf_learner_v2
-description: "A PDF learning application."
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-# In Windows, build-name is used as the major, minor, and patch parts
-# of the product and file versions while build-number is used as the build suffix.
+description: PDF 문서 학습 및 관리 도구
+publish_to: 'none'
 version: 1.0.0+1
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  # UI
+  cupertino_icons: ^1.0.5
   
-  # UI 관련
-  google_fonts: ^6.1.0
-  flutter_svg: ^2.0.9
+  # 환경설정
+  flutter_dotenv: ^5.1.0
+  
+  # 상태관리
+  provider: ^6.0.5
+  
+  # PDF 관련
   syncfusion_flutter_pdfviewer: ^29.1.33
-  shimmer: ^3.0.0
-  
-  # 상태 관리
-  provider: ^6.1.1
-  
-  # 네트워크 및 저장소
-  http: ^1.1.0
-  shared_preferences: ^2.2.2
-  path_provider: ^2.1.1
+  path_provider: ^2.0.15
+  file_picker: ^9.2.2
+  pdf_render: ^1.4.0
+  flutter_pdfview: ^1.3.0
   
   # 유틸리티
   intl: ^0.19.0
-  uuid: ^4.2.1
-  path: ^1.8.3
-  cross_file: ^0.3.3+7
-  mime: ^1.0.4
-  
-  # 추가 유틸리티
-  google_sign_in: ^6.2.1
-  sign_in_with_apple: ^5.0.0
-  share_plus: ^7.2.1
-  permission_handler: ^11.1.0
-  package_info_plus: ^4.2.0
-  flutter_secure_storage: ^9.0.0
-  flutter_adaptive_scaffold: ^0.1.7
-  universal_html: ^2.2.4
+  shared_preferences: ^2.2.0
+  cross_file: ^0.3.3+6
+  desktop_drop: ^0.5.0
+  google_fonts: ^6.2.1
+  flutter_svg: ^2.0.10+1
+  url_launcher: ^6.2.4
+  http: ^1.2.0
+  js: ^0.7.2
+
+  # 추가 필요한 패키지들
+  firebase_core: ^3.12.1
+  firebase_auth: ^5.5.1
+  cloud_firestore: ^5.6.5
+  firebase_storage: ^12.4.4
+  google_sign_in: ^6.1.4
+  get_it: ^8.0.3
+  injectable: ^2.1.2
+  uuid: ^4.5.1
   syncfusion_flutter_pdf: ^29.1.33
-  
-  # Firebase 관련
-  firebase_core: ^2.21.0
-  firebase_auth: ^4.12.1
-  cloud_firestore: ^4.12.2
-  firebase_storage: ^11.4.1
-  firebase_analytics: ^10.6.3
-  firebase_crashlytics: ^3.4.3
-  firebase_performance: ^0.9.3+3
-  firebase_remote_config: ^4.3.3
-  
-  # 환경 설정
-  flutter_dotenv: ^5.1.0
-  
-  # 의존성 주입
-  get_it: ^7.6.4
-  injectable: ^2.3.2
-  
-  # 웹 관련
-  webview_flutter: ^4.4.2
-  url_launcher: ^6.2.1
+  sign_in_with_apple: ^6.1.4
+  share_plus: ^10.1.4
+  permission_handler: ^11.2.0
+  package_info_plus: ^8.3.0
+  webview_flutter: ^4.5.0
+  flutter_secure_storage: ^10.0.0-beta.4
+  image_picker: ^1.0.7
+  mime: ^2.0.0
+  universal_html: ^2.2.4
+  json_annotation: ^4.9.0
+  equatable: ^2.0.7
+  path: ^1.9.1
+  firebase_analytics: ^11.4.4
+  firebase_crashlytics: ^4.3.4
+  firebase_performance: ^0.10.1+4
+  firebase_remote_config: ^5.4.2
+  flutter_local_notifications: ^19.0.0
+  google_mobile_ads: ^5.3.1
+  in_app_purchase: ^3.1.13
+  device_info_plus: ^11.3.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^5.0.0
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^2.0.0
-  build_runner: ^2.4.6
-  injectable_generator: ^2.4.1
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
+  
   assets:
-    - assets/images/
-    - assets/icons/
-    - assets/lang/
-    - assets/sample/
     - .env
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages 
+    - assets/images/
+    - assets/icons/ 
